[CIR][Lowering] Fix the result type of castop lowering
For int_to_bool we always convert to cir.bool and then rely on type
conversion to get us to an i8 and some truncs/zexts to handle LLVM
treating a C++ bool as an i1 at times and an i8 at others. I mistakenly
used the srcType here.

A test for this is in loopop lowering as follows.