[CIR][CIRGen] Pass field index to cir.struct_element_addr
A `member_index` index attribute in `cir.struct_element_addr` now holds
the index for the member being accessed. A APInt is used as the storage
type for the index, alongside a custom builder to abstract the APInt
object creation.

Before, we only passed the name of the field to cir.struct_element_addr,
which was not very useful since it couldn't be used to recover the index
of the member being accessed. This index is essential for lowering CIR
to LLVM, as LLVM needs to know which element is being accessed in the
struct data aggregate.

ghstack-source-id: fa68727f9bd65016647d7ae3acbbf1b109fa79ec
Pull Request resolved: https://github.com/llvm/clangir/pull/148