[CIR][IR] Refactor parsing/printing of implicitly terminated regions
The `shouldPrintTerm` and `checkBlockTerminator` were replaced in favor
of `omitRegionTerm` and `ensureRegionTerm` respectively. The first is
essentially the same method but simplified. The latter was refactored to
do only two things: check if the terminator omission of a region is
valid and, if so, insert the omitted terminator into the region.

The simplifications mostly leverage the fact that we only omit empty
yield values in a single-block region.

ghstack-source-id: 7b943719ca0fb4ac2d1d29775d7545787c23bcbf
Pull Request resolved: https://github.com/llvm/clangir/pull/321