[CIR][CodeGen] Inline assembly: adds operands (#465)
The next step in inline-assembly support: we add instruction operands!
Nothing interesting, just some copy-pasta from the `codegen` with some
sort of simplifications for now.

Well, I'm not sure `functional-type` is the best way to print operands
though it's used in mlir's `InlineAsmOp`. But anyways, maybe you have a
better idea.

There are two or three steps ahead, so we are not that far from being
able to run something!